Revert "Refactoring line prepare function (#346)"
This reverts commit 4c57f5a8e21e25417c6852f2028307506d3747bb.
