+++ src/server/modes/charts/plugins/datalens/d3/utils/index.ts
@@ -1,6 +1,13 @@
 import {ChartKitWidgetTitle} from '@gravity-ui/chartkit';
 
-import {ServerCommonSharedExtraSettings} from '../../../../../../../shared';
+import {
+    ServerCommonSharedExtraSettings,
+    ServerField,
+    ServerPlaceholderSettings,
+    getAxisMode,
+    isDateField,
+    isNumberField,
+} from '../../../../../../../shared';
 
 export function getChartTitle(
     settings?: ServerCommonSharedExtraSettings,
@@ -9,6 +16,26 @@ export function getChartTitle(
         return {
             text: settings.title,
         };
+    }
+
+    return undefined;
+}
+
+export function getAxisType(field?: ServerField, settings?: ServerPlaceholderSettings) {
+    const axisMode = getAxisMode(settings, field?.guid);
+
+    if (axisMode !== 'discrete') {
+        if (isDateField(field)) {
+            return 'datetime';
+        }
+
+        if (isNumberField(field)) {
+            return 'linear';
+        }
+    }
+
+    if (field) {
+        return 'category';
     }
 
     return undefined;
