+++ src/server/modes/charts/plugins/datalens/highcharts.ts
@@ -1,12 +1,9 @@
 import escape from 'lodash/escape';
 
 import {
-    AxisLabelFormatMode,
-    ChartkitHandlers,
     DATASET_FIELD_TYPES,
     Feature,
     LabelsPositions,
-    LegendDisplayMode,
     PlaceholderId,
     ServerChartsConfig,
     ServerPlaceholder,
@@ -16,6 +13,7 @@ import {
     isDateField,
     isEnabledServerFeature,
 } from '../../../../../shared';
+import {ChartkitHandlers} from '../../../../../shared/constants/chartkit-handlers';
 import {registry} from '../../../../registry';
 
 import {IgnoreProps, applyPlaceholderSettingsToAxis} from './utils/axis-helpers';
@@ -210,7 +208,7 @@ export default (...options: [{shared: ServerChartsConfig} | ServerChartsConfig])
                 if (
                     !isNumericalDataType(yItem.data_type as DATASET_FIELD_TYPES) &&
                     !isDateField(yItem) &&
-                    yPlaceholder.settings?.axisFormatMode !== AxisLabelFormatMode.ByField
+                    yPlaceholder.settings?.axisFormatMode !== 'by-field'
                 ) {
                     // A special formatter that returns text labels on the Y axis
                     yAxis.labels.formatter = function () {
@@ -244,7 +242,7 @@ export default (...options: [{shared: ServerChartsConfig} | ServerChartsConfig])
         chart.zoomType = undefined;
     }
 
-    if (shared.extraSettings?.legendMode === LegendDisplayMode.Hide) {
+    if (shared.extraSettings && shared.extraSettings?.legendMode === 'hide') {
         legend = {
             enabled: false,
         };
