+++ src/server/modes/charts/plugins/datalens/js/js.ts
@@ -25,13 +25,12 @@ import {getDatasetIdAndLayerIdFromKey, getFieldList} from '../../helpers/misc';
 import prepareBackendPivotTableData from '../preparers/backend-pivot-table';
 import {PivotData} from '../preparers/backend-pivot-table/types';
 import {prepareD3BarX, prepareHighchartsBarX} from '../preparers/bar-x';
-import {prepareHighchartsBarY} from '../preparers/bar-y';
 import prepareFlatTableData from '../preparers/flat-table';
 import prepareGeopointData from '../preparers/geopoint';
 import prepareGeopointWithClusterData from '../preparers/geopoint-with-cluster';
 import prepareGeopolygonData from '../preparers/geopolygon';
 import prepareHeatmapData from '../preparers/heatmap';
-import {prepareHighchartsLine} from '../preparers/line';
+import prepareLineData from '../preparers/line';
 import prepareMetricData from '../preparers/metric';
 import preparePivotTableData from '../preparers/old-pivot-table/old-pivot-table';
 import {prepareD3Pie, prepareHighchartsPie} from '../preparers/pie';
@@ -486,20 +485,15 @@ function prepareSingleResult({
     switch (visualization.id) {
         case 'line':
         case 'area':
-        case 'area100p': {
+        case 'area100p':
+        case 'bar':
+        case 'bar100p': {
             if (visualization.id === 'line') {
                 shapes = shared.shapes || [];
                 shapesConfig = shared.shapesConfig;
             }
 
-            prepare = prepareHighchartsLine;
-            rowsLimit = 75000;
-            break;
-        }
-
-        case 'bar':
-        case 'bar100p': {
-            prepare = prepareHighchartsBarY;
+            prepare = prepareLineData;
             rowsLimit = 75000;
             break;
         }
