+++ src/server/modes/charts/plugins/datalens/preparers/bar-x/__tests__/mocks/prepareHighchartsBarX.mock.ts
@@ -1,5 +1,3 @@
-import {AxisMode} from '../../../../../../../../../shared';
-
 const chartEditorMock = {
     getLang: () => {
         return 'en';
@@ -61,7 +59,7 @@ export const barXColoringBaseArgs = {
                 labelsView: 'auto',
                 holidays: 'off',
                 axisFormatMode: 'auto',
-                axisModeMap: {'e83fb454-ca9a-4083-95be-1994e9957655': AxisMode.Continuous},
+                axisModeMap: {'e83fb454-ca9a-4083-95be-1994e9957655': 'continuous'},
             },
         },
         {
@@ -638,7 +636,7 @@ export const barXPrepareForQLArgs = {
                 labelsView: 'auto',
                 holidays: 'off',
                 axisFormatMode: 'auto',
-                axisModeMap: {'built_year-0': AxisMode.Discrete},
+                axisModeMap: {'built_year-0': 'discrete'},
                 disableAxisMode: false,
             },
         },
