+++ src/server/modes/charts/plugins/datalens/preparers/bar-x/highcharts.ts
@@ -1,30 +1,31 @@
 import _isEmpty from 'lodash/isEmpty';
 
 import {
-    AxisLabelFormatMode,
-    AxisMode,
     ChartkitHandlers,
     MINIMUM_FRACTION_DIGITS,
     PlaceholderId,
     ServerPlaceholder,
+    Shared,
     WizardVisualizationId,
     getAxisMode,
     getFakeTitleOrTitle,
     getIsNavigatorEnabled,
     isDateField,
+    isDimensionField,
     isMeasureNameOrValue,
+    isVisualizationWithLayers,
 } from '../../../../../../../shared';
+import {getGradientStops} from '../../utils/color-helpers';
 import {getFieldExportingOptions} from '../../utils/export-helpers';
-import {isLegendEnabled, isNumericalDataType} from '../../utils/misc-helpers';
+import {isGradientMode, isNumericalDataType} from '../../utils/misc-helpers';
 import {
-    getHighchartsColorAxis,
-    isXAxisReversed,
-    shouldUseGradientLegend,
-} from '../helpers/highcharts';
-import {getYPlaceholders} from '../helpers/layers';
-import {getSegmentMap} from '../helpers/segments';
-import {getSegmentsYAxis} from '../line/helpers';
+    getSegmentsIndexInOrder,
+    getSegmentsMap,
+    getSegmentsYAxis,
+    getSortedSegmentsList,
+} from '../line/helpers';
 import {getAxisFormattingByField} from '../line/helpers/axis/getAxisFormattingByField';
+import {getLayerPlaceholderWithItems} from '../line/helpers/axis/getLayerPlaceholderWithItems';
 import {PrepareFunctionArgs} from '../types';
 
 import {prepareBarX} from './prepareBarX';
@@ -34,13 +35,18 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
     const {
         ChartEditor,
         placeholders,
+        resultData,
         colors,
         colorsConfig,
         sort,
+        idToTitle,
         idToDataType,
         visualizationId,
         shared,
+        layerSettings,
+        segments,
     } = args;
+    const {data, order} = resultData;
     const preparedData = prepareBarX(args);
     const {graphs} = preparedData;
 
@@ -50,14 +56,41 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
     const xIsNumber = Boolean(xDataType && isNumericalDataType(xDataType));
     const xIsFloat = x ? xDataType === 'float' : null;
     const xIsDate = Boolean(xDataType && isDateField({data_type: xDataType}));
+
     const xPlaceholderSettings = xPlaceholder?.settings;
     const xAxisMode = getAxisMode(xPlaceholderSettings, x?.guid);
-    const isXDiscrete = xAxisMode === AxisMode.Discrete;
+    const isXDiscrete = xAxisMode === 'discrete';
+
     const x2 = placeholders[0].items[1];
+
     const yPlaceholder = placeholders.find((p) => p.id === PlaceholderId.Y);
     const yFields = yPlaceholder?.items || [];
+
     const colorItem = colors[0];
-    const segmentsMap = getSegmentMap(args);
+    const colorFieldDataType = colorItem ? idToDataType[colorItem.guid] : null;
+
+    const gradientMode =
+        colorItem &&
+        colorFieldDataType &&
+        isGradientMode({colorField: colorItem, colorFieldDataType, colorsConfig});
+
+    const sortItem = sort?.[0];
+    const isSortItemExists = Boolean(sort && sort.length);
+    const sortXItem = sort.find((s) => x && s.guid === x.guid);
+
+    const segmentField = segments[0];
+    const segmentIndexInOrder = getSegmentsIndexInOrder(order, segmentField, idToTitle);
+    const segmentsList = getSortedSegmentsList({
+        sortItem,
+        segmentField,
+        segmentIndexInOrder,
+        data,
+        idToDataType,
+    });
+    const segmentsMap = getSegmentsMap({
+        segments: segmentsList,
+        y2SectionItems: [],
+    });
     const isSegmentsExists = !_isEmpty(segmentsMap);
 
     // Here we manage the highcharts settings depending on the parameters
@@ -72,11 +105,7 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
 
             const customConfig: any = {
                 xAxis: {},
-                plotOptions: {
-                    column: {
-                        borderWidth: 1,
-                    },
-                },
+                plotOptions: {},
                 axesFormatting: {xAxis: [], yAxis: []},
                 exporting: {
                     csv: {
@@ -88,16 +117,46 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
                 },
             };
 
-            if (xPlaceholder?.settings?.axisFormatMode === AxisLabelFormatMode.ByField) {
+            if (xPlaceholder?.settings?.axisFormatMode === 'by-field') {
                 customConfig.axesFormatting.xAxis.push(
                     getAxisFormattingByField(xPlaceholder, visualizationId),
                 );
             }
 
-            const [layerYPlaceholder, layerY2Placeholder] = getYPlaceholders(args);
+            const visualization = shared.visualization as Shared['visualization'];
+
+            let layerYPlaceholder;
+            let layerY2Placeholder;
+
+            if (isVisualizationWithLayers(visualization)) {
+                const lastLayer = visualization.layers[visualization.layers.length - 1];
+
+                if (lastLayer.layerSettings.id === layerSettings.id) {
+                    layerYPlaceholder = yPlaceholder;
+                    layerY2Placeholder = undefined;
+
+                    if (!layerYPlaceholder || !layerYPlaceholder.items.length) {
+                        layerYPlaceholder = getLayerPlaceholderWithItems(
+                            shared.visualization,
+                            PlaceholderId.Y,
+                            {isFirstFromTheTop: true},
+                        );
+                    }
+
+                    layerY2Placeholder = getLayerPlaceholderWithItems(
+                        shared.visualization,
+                        PlaceholderId.Y2,
+                        {isFirstFromTheTop: true},
+                    );
+                }
+            } else {
+                layerYPlaceholder = yPlaceholder;
+                layerY2Placeholder = undefined;
+            }
 
             if (
-                layerYPlaceholder?.settings?.axisFormatMode === AxisLabelFormatMode.ByField &&
+                layerYPlaceholder &&
+                layerYPlaceholder.settings?.axisFormatMode === 'by-field' &&
                 !isSegmentsExists
             ) {
                 customConfig.axesFormatting.yAxis.push(
@@ -106,7 +165,8 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
             }
 
             if (
-                layerY2Placeholder?.settings?.axisFormatMode === AxisLabelFormatMode.ByField &&
+                layerY2Placeholder &&
+                layerY2Placeholder.settings?.axisFormatMode === 'by-field' &&
                 !isSegmentsExists
             ) {
                 customConfig.axesFormatting.yAxis.push(
@@ -114,23 +174,70 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
                 );
             }
 
-            if (shouldUseGradientLegend(colorItem, colorsConfig, shared)) {
-                customConfig.colorAxis = getHighchartsColorAxis(graphs, colorsConfig);
+            const isLegendEnabled = shared.extraSettings?.legendMode !== 'hide';
+
+            const isCombinedChartColorizedBySomeDimension =
+                shared.visualization.id === 'combined-chart' &&
+                shared.visualization.layers?.some((layer) => {
+                    return layer.commonPlaceholders.colors.some((field) => isDimensionField(field));
+                });
+
+            const isShouldShowMeasureLegend =
+                gradientMode && !isCombinedChartColorizedBySomeDimension;
+
+            if (isShouldShowMeasureLegend) {
+                const points: Highcharts.PointOptionsObject[] = (graphs as any[]).reduce(
+                    (acc: Highcharts.PointOptionsObject[], graph) => [...acc, ...graph.data],
+                    [],
+                );
+                const colorValues = points
+                    .map((point) => point.colorValue)
+                    .filter((cv): cv is number => Boolean(cv));
+
+                const minColorValue = Math.min(...colorValues);
+                const maxColorValue = Math.max(...colorValues);
+
+                customConfig.colorAxis = {
+                    startOnTick: false,
+                    endOnTick: false,
+                    min: minColorValue,
+                    max: maxColorValue,
+                    stops: getGradientStops(colorsConfig, points, minColorValue, maxColorValue),
+                };
                 customConfig.legend = {
                     title: {
                         text: getFakeTitleOrTitle(colorItem),
                     },
-                    enabled: isLegendEnabled(shared.extraSettings),
+                    enabled: isLegendEnabled,
                     symbolWidth: null,
+                };
+                customConfig.plotOptions = {
+                    bar: {
+                        borderWidth: 1,
+                    },
+                    column: {
+                        borderWidth: 1,
+                    },
                 };
             }
 
             if (xIsDate || xIsNumber) {
-                customConfig.xAxis.reversed = isXAxisReversed(
-                    x,
-                    sort,
-                    visualizationId as WizardVisualizationId,
-                );
+                if (
+                    isSortItemExists &&
+                    sortXItem &&
+                    (sortXItem.direction === 'DESC' || !sortXItem.direction)
+                ) {
+                    if (
+                        visualizationId === WizardVisualizationId.Bar ||
+                        visualizationId === WizardVisualizationId.Bar100p
+                    ) {
+                        // It turns out that in order to expand the X-axis for a Bar chart in Highcharts, you need to pass false
+                        // While in all other types of charts you need to pass true
+                        customConfig.xAxis.reversed = false;
+                    } else {
+                        customConfig.xAxis.reversed = true;
+                    }
+                }
 
                 if (
                     isXDiscrete &&
@@ -177,9 +284,7 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
 
             if (isSegmentsExists) {
                 customConfig.legend = {
-                    enabled:
-                        Boolean(colorItem || x2 || yFields.length > 1) &&
-                        isLegendEnabled(shared.extraSettings),
+                    enabled: Boolean(colorItem || x2 || yFields.length > 1) && isLegendEnabled,
                 };
 
                 const {yAxisFormattings, yAxisSettings} = getSegmentsYAxis(
@@ -199,14 +304,16 @@ export function prepareHighchartsBarX(args: PrepareFunctionArgs) {
             const customConfig: any = {xAxis: {}};
 
             if (xIsDate || xIsNumber) {
-                customConfig.xAxis.reversed = isXAxisReversed(
-                    x,
-                    sort,
-                    visualizationId as WizardVisualizationId,
-                );
+                if (
+                    isSortItemExists &&
+                    sortXItem &&
+                    (sortXItem.direction === 'DESC' || !sortXItem.direction)
+                ) {
+                    customConfig.xAxis.reversed = true;
+                }
 
                 if (xIsDate) {
-                    if (xAxisMode === AxisMode.Discrete) {
+                    if (xAxisMode === 'discrete') {
                         customConfig.xAxis.labels = {
                             formatter: ChartkitHandlers.WizardXAxisFormatter,
                         };
