+++ src/server/modes/charts/plugins/datalens/preparers/bar-x/prepareBarX.ts
@@ -1,7 +1,6 @@
 import _isEmpty from 'lodash/isEmpty';
 
 import {
-    AxisMode,
     HighchartsSeriesCustomObject,
     PlaceholderId,
     ServerField,
@@ -27,10 +26,11 @@ import {
     numericCollator,
 } from '../../utils/misc-helpers';
 import {addActionParamValue} from '../helpers/action-params';
-import {getSegmentMap} from '../helpers/segments';
 import {
     getSegmentsIndexInOrder,
+    getSegmentsMap,
     getSortedCategories,
+    getSortedSegmentsList,
     getXAxisValue,
     prepareLines,
 } from '../line/helpers';
@@ -108,7 +108,17 @@ export function prepareBarX(args: PrepareFunctionArgs) {
 
     const segmentField = segments[0];
     const segmentIndexInOrder = getSegmentsIndexInOrder(order, segmentField, idToTitle);
-    const segmentsMap = getSegmentMap(args);
+    const segmentsList = getSortedSegmentsList({
+        sortItem,
+        segmentField,
+        segmentIndexInOrder,
+        data,
+        idToDataType,
+    });
+    const segmentsMap = getSegmentsMap({
+        segments: segmentsList,
+        y2SectionItems: [],
+    });
     const isSegmentsExists = !_isEmpty(segmentsMap);
 
     const isShapeItemExist = false;
@@ -286,7 +296,7 @@ export function prepareBarX(args: PrepareFunctionArgs) {
         const graphs: any[] = [];
         const uniqueTitles: string[] = [];
 
-        const isXDiscrete = xAxisMode === AxisMode.Discrete;
+        const isXDiscrete = xAxisMode === 'discrete';
         const isSortNumberTypeXAxisByMeasure =
             isSortCategoriesAvailable &&
             isSortItemExists &&
@@ -528,7 +538,7 @@ export function prepareBarX(args: PrepareFunctionArgs) {
         ];
 
         // If there are dates on the X axis, then we pass them as dates
-        if (xIsDate && xAxisMode !== AxisMode.Discrete) {
+        if (xIsDate && xAxisMode !== 'discrete') {
             return {graphs, categories_ms: categories};
         } else {
             return {graphs, categories};
