--- src/server/modes/charts/plugins/datalens/preparers/bar-y/index.ts
@@ -1,1 +1,0 @@
-export {prepareHighchartsBarY} from './highcharts';
