+++ src/server/modes/charts/plugins/datalens/preparers/line-time/__tests__/mocks/line-time.mock.ts
@@ -1,5 +1,3 @@
-import {AxisMode} from '../../../../../../../../../shared';
-
 export const expectedResult = {
     timeline: [
         new Date('2023-03-01T00:00:00').getTime() / 1000,
@@ -83,7 +81,7 @@ export const options = {
                 holidays: 'off',
                 axisFormatMode: 'auto',
                 axisModeMap: {},
-                axisMode: AxisMode.Continuous,
+                axisMode: 'continuous',
             },
         },
         {
@@ -330,7 +328,7 @@ export const options = {
                         holidays: 'off',
                         axisFormatMode: 'auto',
                         axisModeMap: {},
-                        axisMode: AxisMode.Continuous,
+                        axisMode: 'continuous',
                     },
                 },
                 {
