+++ src/server/modes/charts/plugins/datalens/preparers/line/__tests__/line.test.ts
@@ -1,4 +1,4 @@
-import {prepareHighchartsLine as linePrepare} from '../index';
+import linePrepare from '../index';
 
 const {
     regularTitle,
