+++ src/server/modes/charts/plugins/datalens/preparers/line/helpers/segments/getSegmentsList.ts
@@ -1,5 +1,5 @@
-import {SortDirection, isNumberField} from '../../../../../../../../../shared';
-import {collator, numericStringCollator} from '../../../../utils/misc-helpers';
+import {SortDirection} from '../../../../../../../../../shared';
+import {collator, isNumericalDataType, numericStringCollator} from '../../../../utils/misc-helpers';
 
 import {getSegmentName} from './getSegmentName';
 import {GetSegmentsList, GetSortedSegmentsList} from './types';
@@ -23,14 +23,17 @@ export const getSegmentsList = (args: GetSegmentsList): string[] => {
 };
 
 export const getSortedSegmentsList = (args: GetSortedSegmentsList) => {
-    const {data, segmentIndexInOrder, sortItem, segmentField} = args;
+    const {data, segmentIndexInOrder, sortItem, segmentField, idToDataType} = args;
 
     if (!segmentField) {
         return [];
     }
 
     const segments = getSegmentsList({data, segmentField, segmentIndexInOrder});
-    const sortFn = isNumberField(segmentField) ? numericStringCollator : collator.compare;
+
+    const segmentType = idToDataType[segmentField.guid] || segmentField.data_type;
+    const sortFn = isNumericalDataType(segmentType) ? numericStringCollator : collator.compare;
+
     segments.sort(sortFn);
 
     if (
