+++ src/server/modes/charts/plugins/datalens/preparers/scatter/d3.ts
@@ -6,7 +6,7 @@ import type {
 
 import {ServerField, getFakeTitleOrTitle} from '../../../../../../../shared';
 import {PointCustomData, ScatterSeriesCustomData} from '../../../../../../../shared/types/chartkit';
-import {getAxisType} from '../helpers/axis';
+import {getAxisType} from '../../d3/utils';
 import {PrepareFunctionArgs} from '../types';
 
 import {ScatterGraph, prepareScatter} from './prepareScatter';
