+++ src/server/modes/charts/plugins/datalens/utils/misc-helpers.ts
@@ -5,10 +5,8 @@ import {
     ColorMode,
     CommonNumberFormattingOptions,
     DATASET_FIELD_TYPES,
-    LegendDisplayMode,
     MINIMUM_FRACTION_DIGITS,
     ServerColor,
-    ServerCommonSharedExtraSettings,
     ServerField,
     ServerPlaceholder,
     ServerPointSizeConfig,
@@ -361,10 +359,6 @@ function isGradientMode({
         isMeasureField(colorField) ||
         (isNumericalDataType(colorFieldDataType) && colorsConfig.colorMode === ColorMode.GRADIENT)
     );
-}
-
-export function isLegendEnabled(chartSetting?: ServerCommonSharedExtraSettings) {
-    return chartSetting?.legendMode !== LegendDisplayMode.Hide;
 }
 
 export {
