+++ src/shared/types/wizard/placeholder.ts
@@ -54,3 +54,5 @@ export type Placeholder = {
     allowedDataTypes?: Set<string>;
     allowedFinalTypes?: Set<string>;
 };
+
+export type AxisMode = 'discrete' | 'continuous';
