+++ src/ui/units/wizard/constants/index.ts
@@ -3,6 +3,7 @@ import {
     CLIENT_SIDE_FIELD_PROPS,
     CONFLICT_TOOLTIPS,
     DATASET_ERRORS,
+    DATE_MODES,
     HIDE_LABEL_MODES,
     ITEM_TYPES,
     PREFIX,
@@ -47,6 +48,7 @@ export {
     ITEM_TYPES,
     CONFLICT_TOOLTIPS,
     DATASET_ERRORS,
+    DATE_MODES,
     HIDE_LABEL_MODES,
     PREFIX,
     PRIMITIVE_DATA_TYPES,
